requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pydantic==2.9.0
pydantic-settings==2.5.2

# Bybit API
pybit==5.8.0

# Google Gemini AI
google-generativeai==0.8.3
google-ai-generativelanguage==0.6.10

# Data Processing & Analysis
pandas==2.2.3
numpy==2.1.3
scipy==1.14.1

# Statistical Analysis
statsmodels==0.14.4
arch==7.1.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.10
redis==5.2.0
sqlalchemy==2.0.36

# Web Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1
websockets==14.1

# Async
aiohttp==3.11.7
aiofiles==24.1.0

# Data Validation
pydantic[email]==2.9.0

# Logging & Monitoring
python-json-logger==3.1.0
prometheus-client==0.21.0

# Utilities
python-dateutil==2.9.0
pytz==2024.2
requests==2.32.3

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0

# Type Checking
mypy==1.13.0
types-redis==4.6.0.20241004
types-requests==2.32.0.20241016

# Development
black==24.10.0
isort==5.13.2
flake8==7.1.1
